Allow using underscores to replace dashes
When using dotted notation, it is impossible to use some endpoints as they contain a hyphen/dash.
If attempted to be used with dotted notation, Python interprets it as a subtraction rather than a single name.

This allows users to replace the dash with an underscore and proxmoxer will change it back when sending requests.